+++ src/main/java/lotto/util/validator/Validator.java
@@ -1,0 +1,5 @@
+package lotto.util.validator;
+
+public interface Validator {
+    public void validate(int input) throws IllegalArgumentException;
+}
